+++ x-pack/qa/evil-tests/src/test/java/org/elasticsearch/xpack/security/authc/kerberos/KerberosTestCase.java
@@ -98,7 +98,7 @@ public abstract class KerberosTestCase extends ESTestCase {
     }
 
     @AfterClass
-    public static void restoreLocale() throws Exception {
+    public static void restoreLocale() {
         if (restoreLocale != null) {
             Locale.setDefault(restoreLocale);
             restoreLocale = null;
